EVO improve error handling

+++ ICles/V0/Core/bin/PC_REMOVE_TABLE.pl
@@ -138,18 +138,11 @@ log_erreur("$table_name est inconnue dans $environnement") if not %table_info_sa
 
 if ($force) {
 	my $file_to_remove=$env_sip->get_sqlite_path($table_name."_HISTO");
-	log_info("suppression du fichier de données : $file_to_remove");
-	unlink $file_to_remove or die "Impossible de supprimer $file_to_remove : $!";
+	if (-e $file_to_remove) {
+		log_info("suppression du fichier de données : $file_to_remove");
+		unlink $file_to_remove or die "Impossible de supprimer $file_to_remove : $!";
+	}
 }
-
-
-#######################
-# mise à jour COLUMN_INFO
-#######################
-
-my $column=$env_sip->open_local_table("COLUMN_INFO");
-
-$column->execute("DELETE FROM COLUMN_INFO WHERE TABLE_NAME='$table_name'");
 
 
 #######################
